Add detailPostView 분리
